# Allow to use a dedicated process for each job