@samsonasik 2023-04-04
Could you fix the phsptan notice https://github.com/rectorphp/rector-src/actions/runs/4613330600/jobs/8155197723?pr=3563?