@yguedidi 2023-04-04
> Could you fix the phsptan notice https://github.com/rectorphp/rector-src/actions/runs/4613330600/jobs/8155197723?pr=3563?

@samsonasik not sure it's worth it for now, as this is a proof of concept, it will not be merged as is for sure, and the implementation will completely change for sure if the idea gets accepted 🙂 
I'd prefer to first validate the idea with you maintainers