@yguedidi 2023-04-05
> when using default $rectorConfig->parallel(), it should be similar at least equal fast, if it require changing job size, then it will be bad for DX.

Of course! I meant a fourth parameter that default to false to keep actual behavior. Updated the description to mention that