@keulinho 2023-04-05
Wow really cool :blue_heart: 

When i wasn't able to run rector on our project in one go (after my performance and memory fixes), i thought about the same approach to keep the memory consumption stable in runs on larger codebases, but didn't have the time to actually try that out. So really cool that you had that idea as well and were able to get it working :+1: 

> It has a cost to start processes!

It's not only the cost to start a process, but under the hood rector and phpstan also use a lot of internal caching and memoization, and with each new process those caches are empty again, that's why it's all about finding the balance between reusing a process or starting a new one.

My initial idea was to not tie the behaviour to the number of files per job, but to configure a `maxMemoryPerProcess` in the `parallel()` configuration, so that after each job was handled it's checked how much memory the process already consumed, if it is over the configured threshold the process will be killed and a new one will be spawned. In theory this should limit starting of new processes when it's only necessary (given that you configure a sensible max memory limit)

So in the case of a machine with 12 cores and 16 GiB memory you can configure a `maxNumberOfProcesses` of 12 and `maxMemoryPerProcess` of `1G`. 
```php
$rectorConfig->parallel(maxNumberOfProcesses: 12, maxMemoryPerProcess: '1G', jobSize: 20);
```
By default you could set `maxMemoryPerProcess` to `0`, meaning the feature is basically off and it works like before, so you have to actively have to opt-in.

What do you think? 
From glimpsing at the code i have no idea how easy or hard it would be to implement, but maybe you can give that idea a try and see how it performs?