@yguedidi 2023-04-05
Thank you for your kind words!

> What do you think?

To be honest I didn't thought about this idea!
So good to have another alternative in the discussion! 👍🏼 

My first impression: usually when you set a max memory limit, it's because you want the memory usage to never go above.
If I understood your idea correctly, each process will actually be allowed to go above it for its last processed file.
An edge case could be that this file be a very big file.
So a bit confusing to me as is.

Maybe it's just a matter of the parameter wording to avoid this confusion. `$newProcessThreshold` maybe?

Next point is that we will need a userland memory usage management, is this a good thing? no strong opinion. me I know about the [Symfony Stopwatch](https://symfony.com/doc/current/performance.html#profiling-with-symfony-stopwatch) component that could be used

Finally, your idea looks like changing the concept of "job": the current is file list based, this new is memory usage based