@yguedidi 2023-04-05
@TomasVotruba @samsonasik to mention a real business case for being able to run Rector on a big project in one go: be able to set up Rector as a automated code review tool on an existing project.
On CI machines it breaks too, so impossible complete a first run that will build the cache for following PRs (that will run on changed files only so it's OK)