@TomasVotruba 2023-04-12
Hi @yguedidi , 

thanks for this feature proposal! There is quite a long discussion to follow, so only share my 2 cents: we mimic the PHPStan parallelism, as it has to match its workflow and is easier to maintain in case of PHPStan architecture changes in past and the future.

Before going into code, I'd like to ask you 3 things:

* make the CI pass, so we can test it in case of merge without any further reviews, delays
* could you share 2-3 bullet points bottle-necks this solves
* what are the before/after numbers on one exact project?

Thank you