@yguedidi 2023-04-12
> make the CI pass, so we can test it in case of merge without any further reviews, delays

I'll make it pass if this alternative way of running in parallel get accepted.
Let's assume it's accepted, how would be the best way to configure Rector to use this alternative instead of the actual one that mimic PHPStan? so I can finalize the PR
I would have thought about configuring it with a new config function in ParallelConfig from #3602, but it got rejected.
A fourth boolean parameter `$useSeparateProcessPerJob` in `->parallel()` as suggested in the description maybe?
There is of course no goal to remove the current way of doing parallelism like PHPStan!

> could you share 2-3 bullet points bottle-necks this solves

from description:
- no more excessive usage of memory
- allow to run Rector fully in one go on big projects

> what are the before/after numbers on one exact project?

will depend on the parallel configuration, so with the following one `->parallel(300, 3, 200)` on rector-src:
- on main, each persistent process exceeded the 2G (between 2G and 2.5G) of memory consumption, with a total execution time of ~4m 30s
- on this branch, memory consumption remain stable with less than 1.5G per process, at the cost of moving to ~5m 10s
 of execution time

with job size of 100 (`->parallel(300, 3, 100)`):
- on main, same memory performance wit execution time ~4m 20s (-10s from previous, less communication with the main process)
- on this branch, each process doesn't exceed 1G with execution time ~5m 30s (+20s from previous, more processes to create)

This alternative way of parallelism better fit big projects configured with a high job size