@keulinho 2023-04-13
:+1: from my side for this proposal, it basically solves the issues described here https://github.com/rectorphp/rector/issues/7806 as well.