@yguedidi 2023-04-23
@TomasVotruba would you accept this if I finalize it by making the alternative way configurable with a 4th optional parameter `$useSeparateProcesses = false` to `->parallel()`?
It may help in https://github.com/rectorphp/rector/discussions/7891 :)